feat(05-01): merge legacy Accordion/Alert/Badge into atom tests; git rm flat legacy files
- Accordion.test.tsx: import fireEvent, add toggle-via-click suite (false→true, toggle twice, size=lg class)
- Alert.test.tsx: add soft/positive variant combo test and default-only base-class test (legacy coverage)
- Badge atom test already at parity — no changes needed
- git rm accordion.test.tsx, Alert.test.tsx, Badge.test.tsx (flat legacy React files)
- All 29 atom tests green; zero @testing-library/react, zero jest. references
